+++ CHANGELOG.md
@@ -1,2 +1,6 @@
 # 更新日志
 
+#### 1.0.1 (2020-5-8)
+
+* 增加输出图片类型参数 ```imageType```（```image/png``` &#124; ```image/jpeg``` &#124; ```image/webp```(Chrome支持)）
+* 增加输出图片质量参数 ```imageType```为```image/jpeg```时生效